+++ integration-tests/Makefile.am
@@ -2,5 +2,7 @@
 # Makefile.am for libstorage/integration-tests
 #
 
-SUBDIRS = partitions filesystems
+SUBDIRS = partitions filesystems md
+
+python_PYTHON = storageitu.py
